Fix related to logged out ticket submitter
Related to ticket 2939 on Qubes and ticket 400081 on nanoHub,
this will detect the condition when a user submitted a ticket
while not logged in and did not provide a username.  When
these tickets are commented on instead of reporting the submitter
was emailed but actually not sending the email this change will
cause the email address used when submitting the ticket to be
sent an email. Subsequent work will need to validate replies
to these emails are handled correctly.

(cherry picked from commit faca84b8b30cc6149d1ab19b9487c1861d54ff78)